Split CLI JSON emission
